--- a/Domain.hpp
+++ b/Domain.hpp
@@ -20,7 +20,7 @@
 class Domain {
   private:
     Logger *Log;
-    mysycl::queue qq;  field *bufL, *bufR; // Tools for the BCex
+    mysycl::queue qq;  field *bufL, *bufR, *rbufL, *rbufR; // Tools for the BCex
     int cartDims_[3], cartPeriodic_[3], cartCoords_[3], bcType_[3];
     field boxMin_[3], boxMax_[3], boxSize_[3]; // Global info, physical
     field locMin_[3], locMax_[3], locSize_[3]; // This rank info, physical