--- a/Physics.cpp
+++ b/Physics.cpp
@@ -116,7 +116,7 @@ SYCL_EXTERNAL void cons2prim(id<1> myId, unsigned n, field_array u, field_array 
 
 //-- Fluxes and characteristic velocities.
 //    IMPORTANT: all local quantities, they have been sampled --> access simply by eg. f[VX]
-void physicalFlux(int dir, Metric &g, field vD[FLD_TOT], field uD[FLD_TOT], field f[FLD_TOT], field vf[2], field vt[2] ){
+SYCL_EXTERNAL void physicalFlux(int dir, Metric &g, field vD[FLD_TOT], field uD[FLD_TOT], field f[FLD_TOT], field vf[2], field vt[2] ){
   field alpha = g.alpha(), betai[3], gCov[9], gCon[9];
   g.beta(betai);
   g.g3DCov(gCov);