Adding OpenSYCL support for AMD GPUs
Changes:
- Enabled AMD support using RDC+OpenSYCL
- Minor modifications to bandwidth-bound kernels to improve performance
- Adding GPU-aware-mpi option for communication

@svtcli These are some changes we discussed a while back that we made to improve performance on AMD GPUs using OpenSYCL. The changes can be roughly grouped as:
- Fixes for CMake for OpenSYCL build with ROCm
   - To get this working we are slightly tweaking the build process to support relocatable-device-code (RDC) to allow for source code and kernels to be defined in separate files.
- Tweaks to `parForUpdateDu` and `parForRK` to isolate bandwidth bound portions in separate kernels. This gives a notable speedup for both AMD and NVidia.
- Adding new MPI_Isend/MPI_Irecv option to `Domain::BCex`. This is only enabled for AMD GPUs, but *should* also work for NVidia GPUs. This was mainly done because at this time OpenMPI/OpenUCX tends to use host buffers for `MPI_Sendrecv_replace` (pretty big impact on performance).

Known issues:
- These changes will only support AMD with a GPU-aware-MPI install (or no MPI). 
- The cmake configure currently requires explicitly stating `--offload-arch=gfx9XX` for AMD GPUs. This seems to be a side effect of mixing RDC and OpenSYCL and I don't have a good workaround. I added a note to this end in the README.

Performance uplift:
The following table was run with MPI disabled for a problem of the size `Mx=My=Mz=384`. FOM is reported in millions of cell updates per second (MCUPS)
| Device | MCUPS (before) | MCUPS (after) | Speedup |
|---|---|---|---|
| A100 40GB | 112 | 140 | 1.25x |
| MI210 | 61 | 124 | 2x |