@svtcli 2023-08-09
Hi @seanofthemillers !
Thank you for the patches, they are very welcome. This is a short comment even before I look into it (although I ofc remember our previous discussion and agree to the points): typically we do not adopt pull requests on this repository, as it is out of sync with our internal dev repository (private). 
We will need to discuss internally what the best way to include your changes would be... this is the first time we face this issue. 
Other than that, no problem, we are very happy of your contribution!
News will follow as the rest of the team is back from summer vacation.
Cheers,