@seanofthemillers 2023-08-09
Hi @svtcli, Sounds good. I figured you had a separate repo and debated sending a tarball/patch vs a PR. You can feel free to grab the diff from this PR and close it. I just wanted to make sure you guys had a copy of what we discussed a while back.