@svtcli 2024-06-05
We have had a good look at the change, and manually selected some for our new release.
We rejected or rewrote some others. Also the code has changed significantly in the meantime, so we won't merge this request as it is, but should some of these ideas still be needed in future, we can discuss it further. Thanks again!